crescer/las-vegas/api/src/main/java/br/com/cwi/crescer/api/service/ListLoggedUserCoursesService.java
package br.com.cwi.crescer.api.service;

import br.com.cwi.crescer.api.controller.response.CourseResponse;
import br.com.cwi.crescer.api.mapper.CourseListMapper;
import br.com.cwi.crescer.api.model.Course;
import br.com.cwi.crescer.api.security.model.UserAccount;
import br.com.cwi.crescer.api.security.service.UserAccountAuthenticatedService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@Transactional
public class ListLoggedUserCoursesService {

    @Autowired
    private UserAccountAuthenticatedService userAccountAuthenticatedService;

    @Autowired
    private CourseListMapper courseMapper;

    public List<CourseResponse> listLoggedUserCourses() {
        UserAccount loggedUser = userAccountAuthenticatedService.get();

        List<Course> courses = loggedUser.getCreatedCourses();

        return courseMapper.toResponseList(courses);
    }}